requirements.txt
appdirs==1.4.0
coreapi==2.2.4
coreschema==0.0.4
Django==1.10.5
django-cors-headers==2.0.2
django-filter==1.0.1
django-guardian==1.4.6
django-timezone-field==2.0
djangorestframework==3.5.4
itypes==1.1.0
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==0.23
packaging==16.8
psycopg2==2.6.2
pycodestyle==2.3.1
Pygments==2.2.0
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
